hw6: Task1: Task2: Task5: Refactoring
