+++ src/main/java/edu/hw6/task1/DiskMap.java
@@ -1,6 +1,6 @@
 package edu.hw6.task1;
 
-import edu.hw6.Generated;
+import edu.common.Generated;
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.FileReader;
@@ -14,6 +14,28 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.jetbrains.annotations.NotNull;
 
+/**
+ * The {@code DiskMap} is a class that implements the {@link Map} interface and provides a persistent
+ * key-value store backed by a disk file. The data is stored in a simple text file
+ * where each line represents a key-value pair in the format "key:value".
+ * <p>
+ * This implementation utilizes a {@link HashMap} in memory to efficiently manage the key-value pairs.
+ * Operations such as put, remove, and clear automatically trigger a corresponding update
+ * to the disk file to maintain consistency between the in-memory map and the persisted data.
+ * <p>
+ * The file path for the disk storage is specified during the construction of the {@link DiskMap} object.
+ * <p>
+ * The class also incorporates logging using Log4j for error reporting during file I/O operations.
+ *
+ * @see Map
+ * @see HashMap
+ * @see BufferedReader
+ * @see BufferedWriter
+ * @see FileReader
+ * @see FileWriter
+ * @see LogManager
+ * @see Logger
+ */
 public class DiskMap implements Map<String, String> {
     private final String filePath;
     private final Map<String, String> map;
@@ -97,6 +119,23 @@ public class DiskMap implements Map<String, String> {
         return map.entrySet();
     }
 
+    /**
+     * Loads key-value pairs from the specified disk file into the in-memory map.
+     * Each line in the file is expected to be in the format "key:value". Lines
+     * that do not conform to this format are ignored.
+     * <p>
+     * If an IOException occurs during the file reading process, an error message
+     * is logged using Log4j, and the exception details are captured in the log.
+     * <p>
+     * This method is automatically called during the construction of the {@link DiskMap}
+     * object to initialize the in-memory map with persisted data from the disk.
+     *
+     * @see BufferedReader
+     * @see FileReader
+     * @see IOException
+     * @see LogManager
+     * @see Logger
+     */
     @Generated
     private void loadFromDisk() {
         try (BufferedReader reader = new BufferedReader(new FileReader(filePath))) {
@@ -114,6 +153,23 @@ public class DiskMap implements Map<String, String> {
         }
     }
 
+    /**
+     * Saves the current key-value pairs from the in-memory map to the specified
+     * disk file. Each entry is written as a line in the format "key:value".
+     * <p>
+     * If an IOException occurs during the file writing process, an error message
+     * is logged using Log4j, and the exception details are captured in the log.
+     * <p>
+     * This method is automatically triggered after operations that modify the
+     * in-memory map, such as put, remove, putAll, and clear, to ensure that changes
+     * are persisted to the disk for durability.
+     *
+     * @see BufferedWriter
+     * @see FileWriter
+     * @see IOException
+     * @see LogManager
+     * @see Logger
+     */
     @Generated
     private void saveToDisk() {
         try (BufferedWriter writer = new BufferedWriter(new FileWriter(filePath))) {
